Breakout virtualDom from harness
