+++ src/main.ts
@@ -1,7 +1,8 @@
-import harness, { assignChildProperties, assignProperties, replaceChild, replaceChildProperties, replaceProperties } from './harness';
+import harness from './harness';
 import ClientErrorCollector from './intern/ClientErrorCollector';
 import assertRender from './support/assertRender';
 import sendEvent from './support/sendEvent';
+import { assignChildProperties, assignProperties, replaceChild, replaceChildProperties, replaceProperties } from './support/virtualDom';
 
 export {
 	assertRender,
